--- src/components/modal/index.js
@@ -1,8 +1,0 @@
-export {
-  default as Modal,
-  Modal as ModalComponent,
-} from '~/components/modal/Modal';
-
-export {
-  default as ConfirmModal,
-} from '~/components/modal/ConfirmModal';
